+++ src/debug/mod.rs
@@ -12,6 +12,11 @@ use crate::vm::ExitCode;
 use anyhow::Result;
 
 pub mod debugger;
+pub mod host_io;
+pub mod breakpoints;
+pub mod readwrite;
+pub mod step;
+pub mod monitor;
 
 /// Copy all bytes of `data` to `buf`.
 /// Return the size of data copied.
